feat(notifications): share hook presentation formatter
