+++ protocol/src/host/notifications/contracts.ts
@@ -2,4 +2,5 @@ import { notificationsSubscribeV10 } from "@traycer/protocol/host/notifications/
 
 export * from "@traycer/protocol/host/notifications/host-notifications";
 export * from "@traycer/protocol/host/notifications/payloads";
+export * from "@traycer/protocol/host/notifications/presentation";
 export { notificationsSubscribeV10 };
